# docs/openapi.quickgo.mvp.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QuickGO MVP Backend API",
    "version": "0.1.0",
    "description": "Machine-readable contract for the implemented QuickGO MVP backend. All paths are served under /api/v1 and use the standard QuickGO response envelope."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development"
    },
    {
      "url": "https://staging-api.quickgo.example/api/v1",
      "description": "Staging"
    },
    {
      "url": "https://api.quickgo.example/api/v1",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "List current customer's orders",
        "operationId": "ordersMine",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create order from active cart",
        "operationId": "ordersCreate",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateOrder" },
        "responses": {
          "201": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/orders/{orderId}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get current customer's order detail",
        "operationId": "ordersDetail",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/OrderId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/orders/{orderId}/cancel": {
      "post": {
        "tags": ["Orders"],
        "summary": "Cancel order before vendor acceptance",
        "operationId": "ordersCancel",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/OrderId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/ReasonOptional" },
        "responses": {
          "201": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/support/tickets": {
      "get": {
        "tags": ["Support"],
        "summary": "List current user's support tickets",
        "operationId": "supportTickets",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Support"],
        "summary": "Create support ticket",
        "operationId": "supportCreateTicket",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateSupportTicket" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
...
    "/admin/dashboard": {
      "get": {
        "tags": ["Admin"],
        "summary": "Admin operations dashboard",
        "operationId": "adminDashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/orders": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all orders for operations",
        "operationId": "adminOrders",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/orders/{orderId}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get order detail for operations",
        "operationId": "adminOrderDetail",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/OrderId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/orders/{orderId}/assign-rider": {
      "post": {
        "tags": ["Admin"],
        "summary": "Assign rider manually with reason",
        "operationId": "adminAssignRider",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/OrderId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/AssignRider" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/orders/{orderId}/reassign-rider": {
      "post": {
        "tags": ["Admin"],
        "summary": "Reassign rider manually with reason",
        "operationId": "adminReassignRider",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/OrderId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/AssignRider" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/orders/{orderId}/cancel": {
      "post": {
        "tags": ["Admin"],
        "summary": "Cancel order with admin reason",
        "operationId": "adminCancelOrder",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/OrderId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/ReasonRequired" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/orders/{orderId}/payment-collected": {
      "post": {
        "tags": ["Admin"],
        "summary": "Record payment collection by operations",
        "operationId": "adminPaymentCollected",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/OrderId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/PaymentCollected" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/payments/{paymentId}/reconcile": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Reconcile payment with admin reason",
        "operationId": "adminReconcilePayment",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PaymentId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/ReconcilePayment" },
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/vendors": {
      "get": {
        "tags": ["Admin"],
        "summary": "List vendors",
        "operationId": "adminVendors",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create vendor and owner role",
        "operationId": "adminCreateVendor",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateVendor" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/vendors/{vendorId}/status": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Update vendor onboarding or operating status with reason",
        "operationId": "adminVendorStatus",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/VendorId" }],
        "requestBody": { "$ref": "#/components/requestBodies/StatusWithReason" },
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/vendors/{vendorId}/compliance-documents": {
      "get": {
        "tags": ["Admin"],
        "summary": "List vendor compliance documents",
        "operationId": "adminVendorComplianceDocuments",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/VendorId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Store vendor compliance document reference",
        "operationId": "adminCreateVendorComplianceDocument",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/VendorId" }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateDocument" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/vendor-compliance-documents/{documentId}/review": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Review vendor compliance document with reason",
        "operationId": "adminReviewVendorComplianceDocument",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/DocumentId" }],
        "requestBody": { "$ref": "#/components/requestBodies/ReviewDocument" },
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/riders": {
      "get": {
        "tags": ["Admin"],
        "summary": "List riders",
        "operationId": "adminRiders",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create rider and role",
        "operationId": "adminCreateRider",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateRider" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/riders/{riderId}/status": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Update rider onboarding or operating status with reason",
        "operationId": "adminRiderStatus",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RiderId" }],
        "requestBody": { "$ref": "#/components/requestBodies/StatusWithReason" },
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/riders/{riderId}/kyc-documents": {
      "get": {
        "tags": ["Admin"],
        "summary": "List rider KYC documents",
        "operationId": "adminRiderKycDocuments",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RiderId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Store rider KYC document reference",
        "operationId": "adminCreateRiderKycDocument",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RiderId" }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateDocument" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/rider-kyc-documents/{documentId}/review": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Review rider KYC document with reason",
        "operationId": "adminReviewRiderKycDocument",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/DocumentId" }],
        "requestBody": { "$ref": "#/components/requestBodies/ReviewDocument" },
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/products": {
      "get": {
        "tags": ["Admin"],
        "summary": "List products",
        "operationId": "adminProducts",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create product for vendor",
        "operationId": "adminCreateProduct",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateProduct" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/products/{productId}/status": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Update product approval or availability status with reason",
        "operationId": "adminProductStatus",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProductId" }],
        "requestBody": { "$ref": "#/components/requestBodies/StatusWithReason" },
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/categories/{categoryId}/status": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Activate or pause category with reason",
        "operationId": "adminCategoryStatus",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CategoryId" }],
        "requestBody": { "$ref": "#/components/requestBodies/CategoryStatus" },
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
...
    "/admin/support-tickets": {
      "get": {
        "tags": ["Admin"],
        "summary": "List support tickets",
        "operationId": "adminSupportTickets",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/support-tickets/{ticketId}": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Update support ticket with reason",
        "operationId": "adminSupportTicketUpdate",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TicketId" }],
        "requestBody": { "$ref": "#/components/requestBodies/SupportTicketUpdate" },
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
...
    "/admin/audit-logs": {
      "get": {
        "tags": ["Admin"],
        "summary": "List audit logs",
        "operationId": "adminAuditLogs",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/payouts": {
      "get": {
        "tags": ["Admin"],
        "summary": "List vendor and rider payout records",
        "operationId": "adminPayouts",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/admin/payouts/{payoutId}/approve": {
      "post": {
        "tags": ["Admin"],
        "summary": "Manually update payout status with reason",
        "operationId": "adminPayoutApprove",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "payoutId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/PayoutUpdate" },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": { "type": "string" }
      },
      "OrderId": {
        "name": "orderId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "VendorId": {
        "name": "vendorId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "RiderId": {
        "name": "riderId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "ProductId": {
        "name": "productId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "CategoryId": {
        "name": "categoryId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
      "PaymentId": {
        "name": "paymentId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
      "DocumentId": {
        "name": "documentId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "TicketId": {
        "name": "ticketId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
    "responses": {
      "Success": {
        "description": "Standard QuickGO success envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/SuccessEnvelope" }
          }
        }
      },
      "Error": {
        "description": "Standard QuickGO error envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      }
    },
    "requestBodies": {
...
      "CreateOrder": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrder" } } }
      },
      "ReasonOptional": {
        "required": false,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReasonOptional" } } }
      },
      "ReasonRequired": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReasonRequired" } } }
      },
...
      "PaymentCollected": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentCollected" } } }
      },
      "CreateSupportTicket": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateSupportTicket" } } }
      },
...
      "AssignRider": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssignRider" } } }
      },
      "ReconcilePayment": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReconcilePayment" } } }
      },
      "CreateVendor": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateVendor" } } }
      },
      "CreateRider": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateRider" } } }
      },
      "CreateProduct": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProduct" } } }
      },
      "StatusWithReason": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StatusWithReason" } } }
      },
      "CategoryStatus": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CategoryStatus" } } }
      },
      "CreateDocument": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateDocument" } } }
      },
      "ReviewDocument": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReviewDocument" } } }
      },
...
      "SupportTicketUpdate": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SupportTicketUpdate" } } }
      },
      "PayoutUpdate": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PayoutUpdate" } } }
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": ["success", "data", "message"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "nullable": true },
          "message": { "type": "string", "example": "OK" }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "VALIDATION_ERROR" },
              "message": { "type": "string" },
              "details": { "nullable": true }
            }
          }
        }
      },
...
      "CreateOrder": {
        "type": "object",
        "required": ["address_id", "payment_method"],
        "properties": {
          "address_id": { "type": "string" },
          "payment_method": { "type": "string", "enum": ["COD", "UPI_ON_DELIVERY"] },
          "customer_note": { "type": "string" }
        }
      },
      "ReasonOptional": {
        "type": "object",
        "properties": { "reason": { "type": "string" } }
      },
      "ReasonRequired": {
        "type": "object",
        "required": ["reason"],
        "properties": { "reason": { "type": "string" } }
      },
...
      "PaymentCollected": {
        "type": "object",
        "required": ["amount"],
        "properties": {
          "amount": { "type": "number" },
          "payment_method_actual": { "type": "string", "enum": ["COD", "UPI_ON_DELIVERY"] },
          "collector_type": { "type": "string", "enum": ["RIDER", "VENDOR", "ADMIN", "QUICKGO_ADMIN"] },
          "collector_id": { "type": "string" },
          "payment_proof_reference": { "type": "string" },
          "note": { "type": "string" }
        }
      },
      "CreateSupportTicket": {
        "type": "object",
        "required": ["subject", "description"],
        "properties": {
          "order_id": { "type": "string" },
          "priority": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"] },
          "subject": { "type": "string" },
          "description": { "type": "string" }
        }
      },
...
      "AssignRider": {
        "type": "object",
        "required": ["rider_id", "reason"],
        "properties": {
          "rider_id": { "type": "string" },
          "reason": { "type": "string" }
        }
      },
      "ReconcilePayment": {
        "type": "object",
        "required": ["reason"],
        "properties": {
          "reason": { "type": "string" },
          "status": { "type": "string", "enum": ["VERIFIED", "SHORT_COLLECTED", "OVER_COLLECTED", "DISPUTED", "SETTLED"] },
          "amount_collected": { "type": "number" }
        }
      },
      "CreateVendor": {
        "type": "object",
        "required": ["shop_name", "owner_name", "owner_phone", "category_code", "service_zone_id", "address_line", "city", "state", "commission_rate"],
        "properties": {
          "shop_name": { "type": "string" },
          "owner_name": { "type": "string" },
          "owner_phone": { "type": "string" },
          "category_code": { "type": "string", "enum": ["RESTAURANT_FOOD", "VEGETABLES", "FRUITS", "DAIRY"] },
          "service_zone_id": { "type": "string" },
          "address_line": { "type": "string" },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "commission_rate": { "type": "number" }
        }
      },
      "CreateRider": {
        "type": "object",
        "required": ["name", "phone", "service_zone_id"],
        "properties": {
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "service_zone_id": { "type": "string" }
        }
      },
      "CreateProduct": {
        "type": "object",
        "required": ["vendor_id", "category_id", "name", "unit", "price"],
        "properties": {
          "vendor_id": { "type": "string" },
          "category_id": { "type": "string" },
          "name": { "type": "string" },
          "unit": { "type": "string" },
          "price": { "type": "number" },
          "description": { "type": "string" },
          "image_url": { "type": "string" }
        }
      },
      "StatusWithReason": {
        "type": "object",
        "required": ["status", "reason"],
        "properties": {
          "status": { "type": "string" },
          "reason": { "type": "string" },
          "fssai_status": { "type": "string" },
          "is_available": { "type": "boolean" }
        }
      },
      "CategoryStatus": {
        "type": "object",
        "required": ["is_active", "reason"],
        "properties": {
          "is_active": { "type": "boolean" },
          "reason": { "type": "string" }
        }
      },
      "CreateDocument": {
        "type": "object",
        "required": ["type", "document_url"],
        "properties": {
          "type": { "type": "string" },
          "document_url": { "type": "string" },
          "expires_at": { "type": "string", "format": "date-time" }
        }
      },
      "ReviewDocument": {
        "type": "object",
        "required": ["status", "reason"],
        "properties": {
          "status": { "type": "string", "enum": ["APPROVED", "REJECTED", "EXPIRED"] },
          "reason": { "type": "string" },
          "fssai_status": { "type": "string" }
        }
      },
...
      "SupportTicketUpdate": {
        "type": "object",
        "required": ["status", "reason"],
        "properties": {
          "status": { "type": "string" },
          "priority": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"] },
          "admin_note": { "type": "string" },
          "reason": { "type": "string" }
        }
      },
...
    }
  }
}